+++ lumicks/pylake/force_calibration/calibration_results.py
@@ -91,12 +91,36 @@ class CalibrationResults(CalibrationPropertiesMixin):
         """Number of fitted samples (-)."""
         return self.ps_data.total_sampled_used
 
-    def plot(self):
-        """Plot the fitted spectrum"""
+    def plot(self, show_excluded=False, show_active_peak=False):
+        """Plot the fitted spectrum
+
+        Parameters
+        ----------
+        show_excluded : bool
+            Show fitting regions excluded from the fit
+        show_active_peak : bool
+            Show active calibration peak when available
+
+        Raises
+        ------
+        ValueError
+            If specifying `show_active_peak=True` for a passive calibration.
+        """
         import matplotlib.pyplot as plt
 
-        self.ps_data.plot(label="Data")
+        if show_active_peak and not hasattr(self.model, "output_power"):
+            raise ValueError(
+                "Requested to plot an active calibration peak while this is not an active "
+                "calibration. Please specify `show_active_peak=False` when plotting passive"
+                "calibration results."
+            )
+
+        self.ps_data.plot(label="Data", show_excluded=show_excluded)
         self.ps_model.plot(label="Model")
+
+        if show_active_peak:
+            self.model.output_power.ps.plot(label="Active peak")
+
         plt.legend()
 
     def plot_spectrum_residual(self):